# docs/swagger/swagger.json
{
    "swagger": "2.0",
    "info": {
        "description": "工業數據採集閘道系統 API 文檔",
        "title": "Go Gateway API",
        "contact": {},
        "version": "1.0"
    },
    "host": "localhost:8080",
    "basePath": "/api",
    "paths": {
...
        "/datalink/preview/stream": {
            "get": {
                "description": "Opens a text/event-stream backed by the runtime value source for the persisted mapping. Failure responses use the typed safe-error envelope before the stream starts; a runtime stream failure is emitted as an SSE error event.",
                "produces": [
                    "text/event-stream"
                ],
                "tags": [
                    "datalink"
                ],
                "summary": "Stream a persisted mapping preview",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Persisted mapping identifier",
                        "name": "mapping_id",
                        "in": "query",
                        "required": true
                    },
                    {
                        "type": "string",
                        "description": "Workspace scope",
                        "name": "workspace_id",
                        "in": "query",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "SSE stream of connected, preview, heartbeat, and error events",
                        "schema": {
                            "type": "string"
                        }
                    },
                    "422": {
                        "description": "Mapping identifier is missing or invalid",
                        "schema": {
                            "$ref": "#/definitions/handlers.APIErrorResponse"
                        }
                    },
                    "503": {
                        "description": "Preview mapping or runtime stream is unavailable",
                        "schema": {
                            "$ref": "#/definitions/handlers.APIErrorResponse"
                        }
                    }
                }
            }
        },
        "/datalink/runtime/status": {
            "get": {
                "description": "Returns runtime collector state and metrics without synthesizing a ready snapshot on failure.",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "datalink"
                ],
                "summary": "Get runtime status",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Device ID",
                        "name": "device_id",
                        "in": "query"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "object",
                            "additionalProperties": true
                        }
                    },
                    "503": {
                        "description": "Runtime snapshot unavailable",
                        "schema": {
                            "$ref": "#/definitions/handlers.APIErrorResponse"
                        }
                    }
                }
            }
        },
        "/datalink/runtime/stream": {
            "get": {
                "description": "Opens an SSE stream for runtime updates. Failure responses use the typed safe-error envelope.",
                "produces": [
                    "text/event-stream"
                ],
                "tags": [
                    "datalink"
                ],
                "summary": "Stream runtime updates",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Device ID",
                        "name": "device_id",
                        "in": "query",
                        "required": true
                    },
                    {
                        "type": "string",
                        "description": "Comma-separated point IDs",
                        "name": "point_ids",
                        "in": "query"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "SSE stream",
                        "schema": {
                            "type": "string"
                        }
                    },
                    "400": {
                        "description": "Runtime device is missing",
                        "schema": {
                            "$ref": "#/definitions/handlers.APIErrorResponse"
                        }
                    },
                    "503": {
                        "description": "Runtime stream unavailable",
                        "schema": {
                            "$ref": "#/definitions/handlers.APIErrorResponse"
                        }
                    }
                }
            }
        },
        "/datalink/settings": {
            "get": {
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "datalink"
                ],
                "summary": "List durable settings",
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "object",
                            "additionalProperties": true
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/handlers.APIErrorResponse"
                        }
                    }
                }
            }
        },
        "/datalink/settings/{key}": {
            "put": {
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "datalink"
                ],
                "summary": "Update durable settings",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Setting key",
                        "name": "key",
                        "in": "path",
                        "required": true
                    },
                    {
                        "description": "Settings value and expected revision",
                        "name": "request",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/handlers.UpdateSettingRequest"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "object",
                            "additionalProperties": true
                        }
                    },
                    "409": {
                        "description": "Conflict",
                        "schema": {
                            "$ref": "#/definitions/handlers.APIErrorResponse"
                        }
                    },
                    "422": {
                        "description": "Unprocessable Entity",
                        "schema": {
                            "$ref": "#/definitions/handlers.APIErrorResponse"
                        }
                    }
                }
            }
        },
        "/datalink/studio-v2/workspace/activate": {
            "post": {
                "description": "Activates the persisted workspace after hydration, save, and revision checks.",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "studio-v2"
                ],
                "summary": "Activate the Studio v2 workspace",
                "parameters": [
                    {
                        "description": "Activation barrier revisions and save state",
                        "name": "request",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/handlers.ActivateWorkspaceRequest"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "object",
                            "additionalProperties": true
                        }
                    },
                    "409": {
                        "description": "Workspace or settings revision conflict",
                        "schema": {
                            "$ref": "#/definitions/handlers.APIErrorResponse"
                        }
                    },
                    "422": {
                        "description": "Activation barrier or readiness is incomplete",
                        "schema": {
                            "$ref": "#/definitions/handlers.APIErrorResponse"
                        }
                    },
                    "500": {
                        "description": "Activation failed",
                        "schema": {
                            "$ref": "#/definitions/handlers.APIErrorResponse"
                        }
                    }
                }
            }
        },
...
        "/datalink/studio-v2/workspace/runtime-context": {
            "get": {
                "description": "Returns persisted devices and setup context for the runtime monitor.",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "datalink"
                ],
                "summary": "Get runtime workspace context",
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/handlers.runtimeWorkspaceContextResponse"
                        }
                    },
                    "404": {
                        "description": "Runtime device not found",
                        "schema": {
                            "$ref": "#/definitions/handlers.APIErrorResponse"
                        }
                    },
                    "503": {
                        "description": "Workspace or runtime snapshot unavailable",
                        "schema": {
                            "$ref": "#/definitions/handlers.APIErrorResponse"
                        }
                    }
                }
            }
        },
...
    },
    "definitions": {
...
        "handlers.APIErrorResponse": {
            "type": "object",
            "properties": {
                "error": {
                    "$ref": "#/definitions/handlers.TypedAPIErrorEnvelope"
                },
                "success": {
                    "type": "boolean",
                    "example": false
                }
            }
        },
        "handlers.ActivateWorkspaceRequest": {
            "type": "object",
            "required": [
                "readiness_token",
                "settings_revision",
                "workspace_revision"
            ],
            "properties": {
                "pending_saves": {
                    "type": "integer"
                },
                "readiness_token": {
                    "type": "string"
                },
                "save_error": {
                    "type": "string"
                },
                "settings_revision": {
                    "type": "string"
                },
                "workspace_revision": {
                    "type": "string"
                }
            }
        },
...
        "handlers.TypedAPIErrorEnvelope": {
            "type": "object",
            "properties": {
                "action": {
                    "type": "string"
                },
                "code": {
                    "type": "string"
                },
                "dirty_state": {
                    "type": "string"
                },
                "message": {
                    "type": "string"
                },
                "operation_id": {
                    "type": "string"
                },
                "request_id": {
                    "type": "string"
                },
                "retryable": {
                    "type": "boolean"
                },
                "settings_revision": {
                    "type": "string"
                },
                "workspace_revision": {
                    "type": "string"
                }
            }
        },
        "handlers.UpdateSettingRequest": {
            "type": "object",
            "properties": {
                "expected_settings_revision": {
                    "type": "string"
                },
                "value": {}
            }
        },
...
        "handlers.runtimeWorkspaceContextResponse": {
            "type": "object",
            "properties": {
                "default_device_id": {
                    "type": "string"
                },
                "devices": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/handlers.runtimeWorkspaceContextDeviceEntry"
                    }
                },
                "setup": {
                    "$ref": "#/definitions/handlers.runtimeWorkspaceSetupContextResponse"
                },
                "workspace_id": {
                    "type": "string"
                }
            }
        },
...
    }
}