# internal/api/router.go
package api

import (
	"context"
	"fmt"
	"net/http"
	"strings"

	"go-gateway/internal/api/handlers"
	"go-gateway/internal/config"
	"go-gateway/internal/datalink/audit"
	"go-gateway/internal/datalink/collector"
	"go-gateway/internal/datalink/dbtarget"
	"go-gateway/internal/datalink/device"
	"go-gateway/internal/datalink/history"
	"go-gateway/internal/datalink/mapping"
	"go-gateway/internal/datalink/measurement"
	"go-gateway/internal/datalink/modbusshare"
	"go-gateway/internal/datalink/point"
	"go-gateway/internal/datalink/pollinggroup"
	"go-gateway/internal/datalink/recordingplan"
	datalinkruntime "go-gateway/internal/datalink/runtime"
	"go-gateway/internal/datalink/settings"
	"go-gateway/internal/datalink/sourcerule"
	"go-gateway/internal/datalink/tag"
	"go-gateway/internal/datalink/workspace"

	_ "go-gateway/docs/swagger" // Swagger docs

	"github.com/gin-gonic/gin"
	swaggerFiles "github.com/swaggo/files"
	ginSwagger "github.com/swaggo/gin-swagger"
)

// DatalinkServices 包含所有 Datalink 相關服務
type DatalinkServices struct {
	Device                *device.Service
	Point                 *point.Service
	Tag                   *tag.Service
	Mapping               *mapping.Service
	PollingGroup          *pollinggroup.Service
	Settings              *settings.Service
	ModbusShare           *modbusshare.Service
	ModbusShareReconciler *modbusshare.Reconciler
	Scheduler             *collector.Scheduler
	Runtime               *datalinkruntime.Service
	DBTarget              *dbtarget.ConnectorService
	DBMapping             *dbtarget.MappingService
	SourceRule            *sourcerule.Service
	Measurement           *measurement.Service
	RecordingPlan         *recordingplan.Service
	History               *history.Service
	Workspace             *workspace.Service
	WriteGroups           *workspace.WriteGroupService
	// WriteGroupDelivery reads durable group delivery status; nil disables the route.
	WriteGroupDelivery handlers.WriteGroupDeliveryReader
	// WriteGroupTestWrite previews and confirms explicit group test writes; nil
	// makes the test-write routes answer that the capability is unavailable.
	WriteGroupTestWrite        handlers.WriteGroupTestWriter
	Audit                      *audit.Service
	ShareRestore               handlers.ShareRestoreBarrier
	RecordingStartShareBarrier func(context.Context, workspace.RecordingStartRequest) error
}

// NewRouter 建立並配置 Gin 路由器
//
// Args:
//   - datalinkServices: Datalink 服務容器
//
// Returns:
//   - 配置好的 Gin Engine 實例
func NewRouter(datalinkServices *DatalinkServices) *gin.Engine {
	cfg := config.Get()

	// 設定 Gin 模式：根據配置決定是否使用 debug 模式
	// 如果 DEBUG=true 或 GIN_MODE=debug，則使用 debug 模式
	// 否則使用 release 模式（不顯示路由註冊資訊）
	if cfg.Debug {
		gin.SetMode(gin.DebugMode)
	} else {
		gin.SetMode(gin.ReleaseMode)
	}

	router := gin.New()

	// 中間件
	router.Use(customLoggerMiddleware()) // 使用自定義日誌中間件，過濾頻繁的 debug API 請求
	router.Use(gin.Recovery())
	router.Use(corsMiddleware(cfg))

	// API 路由群組（使用配置中的路徑）
	apiV1 := router.Group(cfg.API.BasePath)
	{
		// WebSocket 端點（保留用於其他用途）
		wsHandler := handlers.NewWebSocketHandler()
		apiV1.GET("/ws", wsHandler.HandleWebSocket)

		// SSE 端點（用於監控數據流）
		sseHandler := handlers.NewSSEHandler()

		// Debug 處理器
		debugHandler := handlers.NewDebugHandler()

		apiV1.GET("/test/monitor/stream", sseHandler.HandleMonitorStream)

		// 共用的 TestHandler 實例
		testHandler := handlers.NewTestHandler(wsHandler, sseHandler, debugHandler)

		// 測試相關 API
		testGroup := apiV1.Group("/test")
		testGroup.POST("/connect", testHandler.Connect)
		testGroup.POST("/disconnect", testHandler.Disconnect)
		testGroup.GET("/status", testHandler.GetStatus)
		testGroup.POST("/read", testHandler.Read)
		testGroup.POST("/write", testHandler.Write)
		testGroup.POST("/batch", testHandler.Batch)
		testGroup.POST("/script", testHandler.ExecuteScript)
		testGroup.GET("/scripts", testHandler.ListScripts)
		testGroup.POST("/scripts", testHandler.SaveScript)
		testGroup.DELETE("/scripts/:id", testHandler.DeleteScript)
		testGroup.POST("/monitor/start", testHandler.StartMonitor)
		testGroup.POST("/monitor/stop", testHandler.StopMonitor)

		// Debug 相關 API（使用共用的 debugHandler）
		debugGroup := apiV1.Group("/debug")
		debugGroup.GET("/packets", debugHandler.GetPackets)
		debugGroup.GET("/logs", debugHandler.GetLogs)
		debugGroup.DELETE("/clear", debugHandler.ClearData)
		debugGroup.POST("/send-raw", debugHandler.SendRaw)
		debugGroup.GET("/analyze/:packetId", debugHandler.AnalyzePacket)

		// 配置管理 API
		configGroup := apiV1.Group("/config")
		{
			configHandler := handlers.NewConfigHandler()
			configGroup.GET("/presets", configHandler.GetPresets)
			configGroup.POST("/presets", configHandler.SavePreset)
			configGroup.DELETE("/presets/:id", configHandler.DeletePreset)
		}

		// 測試模板 API
		templateGroup := apiV1.Group("/templates")
		{
			templateHandler := handlers.NewTemplateHandler()
			templateGroup.GET("", templateHandler.List)
			templateGroup.POST("", templateHandler.Save)
			templateGroup.GET("/:id", templateHandler.Get)
			templateGroup.DELETE("/:id", templateHandler.Delete)
		}

		// 連線池 API
		connectionGroup := apiV1.Group("/connections")
		{
			// 使用共用的 testHandler
			connectionHandler := handlers.NewConnectionHandler(testHandler)
			connectionGroup.GET("", connectionHandler.List)
			connectionGroup.GET("/:id", connectionHandler.Get)
		}

		// Datalink API Group
		datalinkGroup := apiV1.Group("/datalink")
		{
			// Health
			healthHandler := handlers.NewDatalinkHealthHandler()
			datalinkGroup.GET("/health", healthHandler.Check)

			// Dashboard
			dashboardHandler := handlers.NewDashboardHandler(
				datalinkServices.Device,
				datalinkServices.Point,
				datalinkServices.Tag,
			)
			datalinkGroup.GET("/dashboard/stats", dashboardHandler.GetStats)
			datalinkGroup.GET("/dashboard/device-statuses", dashboardHandler.GetDeviceStatuses)

			runtimeHandler := handlers.NewRuntimeHandler(
				datalinkServices.Device,
				datalinkServices.Point,
				datalinkServices.PollingGroup,
				datalinkServices.Scheduler,
				datalinkServices.Runtime,
				datalinkServices.Workspace,
			).WithSetupContextServices(
				datalinkServices.SourceRule,
				datalinkServices.Tag,
				datalinkServices.Mapping,
				datalinkServices.DBTarget,
				datalinkServices.DBMapping,
			)
			datalinkGroup.GET("/runtime/status", runtimeHandler.Status)
			runtimeStreamHandler := handlers.NewRuntimeStreamHandler(datalinkServices.Runtime)
			datalinkGroup.GET("/runtime/stream", runtimeStreamHandler.Stream)

			// Protocols
			protocolHandler := handlers.NewProtocolHandler()
			datalinkGroup.GET("/protocols", protocolHandler.List)

			// Polling Groups
			pollingGroupHandler := handlers.NewPollingGroupHandler(datalinkServices.PollingGroup)
			datalinkGroup.GET("/polling-groups", pollingGroupHandler.List)
			datalinkGroup.POST("/polling-groups", pollingGroupHandler.Create)
			datalinkGroup.GET("/polling-groups/:id", pollingGroupHandler.Get)
			datalinkGroup.PUT("/polling-groups/:id", pollingGroupHandler.Update)
			datalinkGroup.DELETE("/polling-groups/:id", pollingGroupHandler.Delete)

			// Devices
			deviceHandler := handlers.NewDeviceHandler(datalinkServices.Device, datalinkServices.Runtime)
			datalinkGroup.GET("/devices", deviceHandler.List)
			datalinkGroup.POST("/devices", deviceHandler.Create)
			datalinkGroup.POST("/devices/test-draft", deviceHandler.TestDraftConnection)
			datalinkGroup.GET("/devices/:id", deviceHandler.Get)
			datalinkGroup.PUT("/devices/:id", deviceHandler.Update)
			datalinkGroup.DELETE("/devices/:id", deviceHandler.Delete)
			datalinkGroup.POST("/devices/:id/test", deviceHandler.TestConnection)
			datalinkGroup.POST("/devices/:id/activate", deviceHandler.Activate)
			datalinkGroup.POST("/devices/:id/disable", deviceHandler.Disable)
			datalinkGroup.POST("/devices/test-batch", deviceHandler.TestConnectionBatch)

			// Source Rules
			if datalinkServices.SourceRule != nil {
				sourceRuleHandler := handlers.NewSourceRuleHandler(datalinkServices.SourceRule)
				datalinkGroup.GET("/source-rules", sourceRuleHandler.List)
				datalinkGroup.POST("/source-rules", sourceRuleHandler.Create)
				datalinkGroup.GET("/source-rules/:id", sourceRuleHandler.Get)
				datalinkGroup.PUT("/source-rules/:id", sourceRuleHandler.Update)
				datalinkGroup.DELETE("/source-rules/:id", sourceRuleHandler.Delete)
				datalinkGroup.POST("/source-rules/:id/enable", sourceRuleHandler.Enable)
				datalinkGroup.POST("/source-rules/:id/disable", sourceRuleHandler.Disable)
				datalinkGroup.GET("/source-rules/:id/candidates", sourceRuleHandler.Candidates)
				datalinkGroup.POST("/source-rules/:id/candidates/recompute", sourceRuleHandler.RecomputeCandidates)
				datalinkGroup.POST("/source-rules/:id/tags/apply", sourceRuleHandler.ApplyTags)
				datalinkGroup.POST("/source-rules/:id/database-outputs/apply", sourceRuleHandler.ApplyDatabaseOutputs)
				datalinkGroup.POST("/source-rules/:id/local-modbus/apply", sourceRuleHandler.ApplyLocalModbusOutputs)
				datalinkGroup.GET("/source-rules/:id/tag-review-decisions", sourceRuleHandler.ListTagReviewDecisions)
				datalinkGroup.POST("/source-rules/:id/tag-review-decisions", sourceRuleHandler.UpsertTagReviewDecision)

				if datalinkServices.Workspace != nil {
					workspaceSourceRuleHandler := handlers.NewStudioV2WorkspaceSourceRulesHandler(datalinkServices.Workspace, datalinkServices.Device, datalinkServices.SourceRule)
					datalinkGroup.GET("/studio-v2/workspace/source-rules", workspaceSourceRuleHandler.List)
					datalinkGroup.POST("/studio-v2/workspace/source-rules", workspaceSourceRuleHandler.Create)
					datalinkGroup.PUT("/studio-v2/workspace/source-rules/:id", workspaceSourceRuleHandler.Update)
					datalinkGroup.DELETE("/studio-v2/workspace/source-rules/:id", workspaceSourceRuleHandler.Delete)
				}
			}

			// Studio V2 Telemetry Recording & History Routes
			registerStudioV2RecordingRoutes(datalinkGroup, datalinkServices)

			// Points
			pointHandler := handlers.NewPointHandler(datalinkServices.Point, datalinkServices.Runtime).
				WithPolling(datalinkServices.Scheduler, datalinkServices.Mapping, datalinkServices.PollingGroup).
				WithDirectReader(handlers.NewPointDirectReader(datalinkServices.Point, datalinkServices.Device))
			datalinkGroup.GET("/points", pointHandler.List)
			datalinkGroup.POST("/points", pointHandler.Create)
			datalinkGroup.POST("/points/batch", pointHandler.BatchCreate)
			datalinkGroup.GET("/points/:id", pointHandler.Get)
			datalinkGroup.PUT("/points/:id", pointHandler.Update)
			datalinkGroup.DELETE("/points/:id", pointHandler.Delete)
			datalinkGroup.POST("/points/:id/poll", pointHandler.Poll)
			datalinkGroup.POST("/points/poll", pointHandler.PollBatch)

			// Tags
			tagHandler := handlers.NewTagHandler(datalinkServices.Tag)
			datalinkGroup.GET("/tags", tagHandler.List)
			datalinkGroup.POST("/tags", tagHandler.Create)
			datalinkGroup.GET("/tags/:id", tagHandler.Get)
			datalinkGroup.PUT("/tags/:id", tagHandler.Update)
			datalinkGroup.DELETE("/tags/:id", tagHandler.Delete)
			datalinkGroup.POST("/tags/:id/activate", tagHandler.Activate)
			datalinkGroup.POST("/tags/:id/retire", tagHandler.Retire)
			datalinkGroup.POST("/tags/batch", tagHandler.BatchCreate)
			datalinkGroup.POST("/tags/validate-key", tagHandler.ValidateKey)

			// Mappings
			mappingHandler := handlers.NewMappingHandler(datalinkServices.Mapping, datalinkServices.Runtime)
			datalinkGroup.GET("/mappings", mappingHandler.List)
			datalinkGroup.POST("/mappings", mappingHandler.Create)
			datalinkGroup.GET("/mappings/:id", mappingHandler.Get)
			datalinkGroup.PUT("/mappings/:id", mappingHandler.Update)
			datalinkGroup.DELETE("/mappings/:id", mappingHandler.Delete)
			datalinkGroup.POST("/mappings/preview", mappingHandler.Preview)
			datalinkGroup.POST("/mappings/validate-pipeline", mappingHandler.ValidatePipeline)
			if datalinkServices.Workspace != nil && datalinkServices.SourceRule != nil {
				workspaceMappingHandler := handlers.NewStudioV2WorkspaceMappingsHandler(
					datalinkServices.Workspace,
					datalinkServices.Device,
					datalinkServices.SourceRule,
					datalinkServices.Point,
					datalinkServices.Tag,
					datalinkServices.Mapping,
				)
				datalinkGroup.GET("/studio-v2/workspace/mappings", workspaceMappingHandler.List)
				datalinkGroup.POST("/studio-v2/workspace/mappings", workspaceMappingHandler.Create)
				datalinkGroup.PUT("/studio-v2/workspace/mappings/:id", workspaceMappingHandler.Update)
				datalinkGroup.DELETE("/studio-v2/workspace/mappings/:id", workspaceMappingHandler.Delete)
			}

			// Settings
			settingsHandler := handlers.NewSettingsHandler(datalinkServices.Settings).WithModbusShare(datalinkServices.ModbusShare)
			datalinkGroup.GET("/settings", settingsHandler.List)
			datalinkGroup.PUT("/settings/:key", settingsHandler.Update)

			if datalinkServices.Workspace != nil {
				workspaceHandler := handlers.NewStudioV2WorkspaceHandler(datalinkServices.Workspace).WithModbusShare(datalinkServices.ModbusShare).WithModbusShareReconciler(datalinkServices.ModbusShareReconciler)
				if builder := canonicalShareDesiredMappingBuilder(datalinkServices); builder != nil {
					workspaceHandler.WithCanonicalPlanBuilder(builder)
				}
				datalinkGroup.GET("/studio-v2/workspace", workspaceHandler.Get)
				if datalinkServices.Audit != nil {
					workspaceAuditHandler := handlers.NewStudioV2WorkspaceAuditHandler(datalinkServices.Workspace, datalinkServices.Audit)
					datalinkGroup.GET("/studio-v2/workspace/audit-history", workspaceAuditHandler.List)
				}
				datalinkGroup.GET("/studio-v2/workspace/runtime-context", runtimeHandler.WorkspaceContext)
				if datalinkServices.Device != nil {
					activationService := workspace.NewActivationService(datalinkServices.Workspace, datalinkServices.Device, datalinkServices.Runtime)
					var workspaceActivationHandler *handlers.StudioV2WorkspaceActivationHandler
					// 若 database 服務齊備，activate 前先確保目標資料表存在（冪等保險）。
					if datalinkServices.DBTarget != nil && datalinkServices.DBMapping != nil && datalinkServices.SourceRule != nil {
						schemaEnsurer := handlers.NewStudioV2WorkspaceDatabaseHandler(
							datalinkServices.Workspace,
							datalinkServices.Device,
							datalinkServices.SourceRule,
							datalinkServices.DBTarget,
							datalinkServices.DBMapping,
						)
						workspaceActivationHandler = handlers.NewStudioV2WorkspaceActivationHandler(activationService, schemaEnsurer)
					} else {
						workspaceActivationHandler = handlers.NewStudioV2WorkspaceActivationHandler(activationService)
					}
					configureModbusShareActivation(workspaceActivationHandler, datalinkServices)
					workspaceActivationHandler.WithAudit(datalinkServices.Audit)
					datalinkGroup.POST("/studio-v2/workspace/activate", workspaceActivationHandler.Activate)

					workspaceDeviceHandler := handlers.NewStudioV2WorkspaceDevicesHandler(datalinkServices.Workspace, datalinkServices.Device, datalinkServices.Runtime)
					datalinkGroup.GET("/studio-v2/workspace/devices", workspaceDeviceHandler.List)
					datalinkGroup.POST("/studio-v2/workspace/devices", workspaceDeviceHandler.Create)
					datalinkGroup.PUT("/studio-v2/workspace/devices/:id", workspaceDeviceHandler.Update)
					datalinkGroup.PUT("/studio-v2/workspace/devices/:id/availability", workspaceDeviceHandler.UpdateAvailability)
					datalinkGroup.DELETE("/studio-v2/workspace/devices/:id", workspaceDeviceHandler.Delete)
					datalinkGroup.PUT("/studio-v2/workspace/device-order", workspaceDeviceHandler.UpdateOrder)
				}
			}

			// SSE Preview Stream
			ssePreviewHandler := handlers.NewDatalinkSSEHandler(datalinkServices.Mapping, datalinkServices.Runtime)
			if datalinkServices.Workspace != nil && datalinkServices.SourceRule != nil && datalinkServices.Point != nil {
				ssePreviewHandler.WithPreviewScopeValidator(sourcerule.NewPreviewMappingScopeChecker(datalinkServices.Workspace, datalinkServices.SourceRule, datalinkServices.Point))
			}
			datalinkGroup.GET("/preview/stream", ssePreviewHandler.PreviewStream)

			// Local Modbus Share
			registerModbusShareRoutes(datalinkGroup, datalinkServices)

			if datalinkServices.DBTarget != nil && datalinkServices.DBMapping != nil {
				dbTargetHandler := handlers.NewDatabaseTargetHandler(
					datalinkServices.DBTarget,
					datalinkServices.DBMapping,
				).WithWriteGroups(datalinkServices.WriteGroups)
				datalinkGroup.GET("/db-targets/connectors", dbTargetHandler.ListConnectors)
				datalinkGroup.POST("/db-targets/connectors", dbTargetHandler.CreateConnector)
				datalinkGroup.GET("/db-targets/connectors/:id", dbTargetHandler.GetConnector)
				datalinkGroup.PUT("/db-targets/connectors/:id", dbTargetHandler.UpdateConnector)
				datalinkGroup.DELETE("/db-targets/connectors/:id", dbTargetHandler.DeleteConnector)
				datalinkGroup.POST("/db-targets/connectors/:id/test", dbTargetHandler.TestConnector)
				datalinkGroup.GET("/db-targets/connectors/:id/tables", dbTargetHandler.ListTables)
				datalinkGroup.GET("/db-targets/connectors/:id/validate", dbTargetHandler.ValidateConnector)
				datalinkGroup.POST("/db-targets/connectors/:id/schema/generate", dbTargetHandler.GenerateSchema)
				datalinkGroup.POST("/db-targets/connectors/:id/mappings/dry-run", dbTargetHandler.DryRunMappings)
				datalinkGroup.GET("/db-targets/connectors/:id/write-history", dbTargetHandler.ListWriteHistory)
				datalinkGroup.GET("/db-targets/mappings", dbTargetHandler.ListMappings)
				datalinkGroup.POST("/db-targets/mappings", dbTargetHandler.CreateMapping)
				datalinkGroup.GET("/db-targets/mappings/:id", dbTargetHandler.GetMapping)
				datalinkGroup.PUT("/db-targets/mappings/:id", dbTargetHandler.UpdateMapping)
				datalinkGroup.DELETE("/db-targets/mappings/:id", dbTargetHandler.DeleteMapping)

				if datalinkServices.Workspace != nil && datalinkServices.SourceRule != nil {
					workspaceDatabaseHandler := handlers.NewStudioV2WorkspaceDatabaseHandler(
						datalinkServices.Workspace,
						datalinkServices.Device,
						datalinkServices.SourceRule,
						datalinkServices.DBTarget,
						datalinkServices.DBMapping,
					).WithAudit(datalinkServices.Audit).WithWriteGroups(datalinkServices.WriteGroups)
					datalinkGroup.GET("/studio-v2/workspace/database-config", workspaceDatabaseHandler.GetConfig)
					datalinkGroup.PUT("/studio-v2/workspace/database-config", workspaceDatabaseHandler.UpdateConfig)
					datalinkGroup.GET("/studio-v2/workspace/database-targets", workspaceDatabaseHandler.ListTargets)
					datalinkGroup.GET("/studio-v2/workspace/database-metadata", workspaceDatabaseHandler.GetMetadata)
					datalinkGroup.PUT("/studio-v2/workspace/database-targets/:point_id", workspaceDatabaseHandler.UpsertTarget)
					datalinkGroup.POST("/studio-v2/workspace/database-schema/generate", workspaceDatabaseHandler.GenerateSchema)
				}
			}
		}
	}

	// Swagger API 文檔
	router.GET("/swagger/*any", ginSwagger.WrapHandler(swaggerFiles.Handler))

	return router
}

// customLoggerMiddleware 自定義日誌中間件，過濾頻繁的 debug API 輪詢請求
func customLoggerMiddleware() gin.HandlerFunc {
	// 配置 Logger，跳過頻繁輪詢的 API 路徑
	// 這些是前端定期輪詢的 API，不需要每次都記錄日誌
	skipPaths := []string{
		"/api/v1/debug/packets",
		"/api/v1/debug/logs",
		"/api/v1/test/status", // 狀態檢查也可能頻繁輪詢
	}

	return gin.LoggerWithConfig(gin.LoggerConfig{
		SkipPaths: skipPaths,
		// 確保所有請求都被記錄（包括 /test/connect）
		Formatter: formatHTTPLog,
	})
}

func formatHTTPLog(param gin.LogFormatterParams) string {
	errorMessage := strings.TrimSpace(param.ErrorMessage)
	if errorMessage != "" {
		errorMessage = " " + errorMessage
	}

	return fmt.Sprintf("[HTTP] %s %-6s %s %3d %s%s\n",
		param.TimeStamp.Format("15:04:05"),
		param.Method,
		param.Path,
		param.StatusCode,
		param.Latency,
		errorMessage,
	)
}

// corsMiddleware 處理 CORS 跨域請求
func corsMiddleware(cfg *config.Config) gin.HandlerFunc {
	return func(c *gin.Context) {
		// 設定允許的來源
		origin := c.Request.Header.Get("Origin")
		allowOrigin := "*"

		// 如果配置了特定的來源列表，檢查是否匹配
		if len(cfg.CORS.AllowOrigins) > 0 && cfg.CORS.AllowOrigins[0] != "*" {
			allowOrigin = ""
			for _, allowed := range cfg.CORS.AllowOrigins {
				if allowed == "*" || allowed == origin {
					allowOrigin = origin
					break
				}
			}
		}

		if allowOrigin != "" {
			c.Writer.Header().Set("Access-Control-Allow-Origin", allowOrigin)
		}
		c.Writer.Header().Set("Access-Control-Allow-Credentials", "true")
		c.Writer.Header().Set("Access-Control-Allow-Headers", "Content-Type, Content-Length, Accept-Encoding, X-CSRF-Token, Authorization, accept, origin, Cache-Control, X-Requested-With")
		c.Writer.Header().Set("Access-Control-Allow-Methods", "POST, OPTIONS, GET, PUT, DELETE")

		if c.Request.Method == "OPTIONS" {
			c.AbortWithStatus(http.StatusNoContent)
			return
		}

		c.Next()
	}
}